# Use `setRange` when copying output chunks to the final buffer in `CodedBufferWriter`