@osa1 2023-10-24
@rakudrama any suggestions on how to improve the dart2js performance here?